Move disk code into its own file
